pyproject.toml
[tool.poetry]
name = "djangoshopproject"
version = "0.1.0"
description = ""
authors = ["Constantine Putilov <putilov.constantine@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
psycopg2 = "^2.9.9"
pillow = "^10.1.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
